refactor(host): let the pairing ceremony take its identities
The ceremony derived the client identity from the connection's certificate. A
browser presents no certificate, so that one line was the whole reason a
WebTransport client could not reuse this code — everything else about SPAKE2
role B is carrier-neutral.

The caller now supplies both identities and the stream types are generic, so the
native plane passes the two certificate fingerprints exactly as before and the
browser plane can pass a key it holds plus the certificate hash it pinned to
connect. Getting them wrong yields a MAC mismatch indistinguishable from a wrong
PIN, which is why the choice belongs to the caller and is spelled out there.

The redundant conversions this drops are fallout from moving the native plane
onto SessionLink: `send_datagram` already takes the owned bytes and `close`
already takes the code.

+++ crates/punktfunk-host/src/webtransport/datagrams.rs
@@ -93,6 +93,6 @@ mod tests {
         assert_eq!(q.len(), INBOX_CAP, "bounded, whatever the sender does");
         // The oldest went, so what is left is the most recent window — the useful end for a
         // live stream, where a stale datagram is worth less than a fresh one.
-        assert_eq!(q.front().map(|d| d[0]), Some(10u8 as u8));
+        assert_eq!(q.front().map(|d| d[0]), Some(10_u8));
     }
 }
